use channel to delete msgs

+++ index.js
@@ -31,7 +31,7 @@ bot.on('message', (message) => {
             || upMsg === '/delete'.toUpperCase()) {
                 message.reply('Deleting... please wait');
                 console.log(message.member);
-                if (message.member.hasPermission("MANAGE_MESSAGES")) {
+                if (!message.webhookID && message.member.hasPermission("MANAGE_MESSAGES")) {
                     message.channel.fetchMessages()
                        .then(function(list){
                             message.channel.bulkDelete(list);
